Separate renderers from generators (#45)
* Definition parser shouldn't create a schema

* Move definition converters to src/converters

* Move schema to ts converters

* Move expressible converters

* Few fixes

* Use schema.python instead of schema.types.python

* Fix compiler errors

* Fix test

* Rename GenerationOutput to Generation

* Init generation module

* Init renderers

* Implement ts renderer

* Write files after rendering generation

* Separate python generator from renderer

* Add playground

* Update gitignore

* Revert empty dir validation

* Remove unused error

* Disallow float literals

* Expression builders separated from model types

* Add individual primitive types

* Cleanup

* Expression builders for python

* Cleanup

* Update snapshots

+++ src/definition/_schemas.ts
@@ -6,7 +6,7 @@ export const primitiveType = z.enum(['nil', 'string', 'boolean', 'int', 'timesta
 
 export const literalType = z.object({
   type: z.literal('literal'),
-  value: z.string().or(z.number()).or(z.boolean()),
+  value: z.string().or(z.number().int()).or(z.boolean()),
 });
 
 export const enumType = z.object({
